Inject the content script into all tabs after extension is newly installed.
This involves changing the initialization workflow a little bit, in that the injection should happen (and finish) first, before we prompt the user to initialize their password.  Please double-check that this should be the case.

Enabling chrome.tabs in the manifest so that we can log the url of the tabs (i.e. make it easier for Wilson to troubleshoot).  If not useful, please let me know to remove this.
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=89911077