Fix re-injection of content scripts whenever the event page (i.e. the non-persistent background page) reloads.  Instead, the content script injection is moved to be triggered once by install event only.
Adding exponential backoff to make sure we wait until the page
initialization are done, i.e. for the managed policies to be set.

Thanks rbyers@, asargent@ for the help here and suggesting the fix.

(1) We do not anticipate the race condition where new tabs/frames are
created after the extension is installed but before the
onInstalledEvent.

(2) Added a TODO to think about how to handle orphaned content scripts
after an update.