Update Password Alert to handle the new login page.
Getting an event on document unload is a bit heavy, but I did not see a better way to get an event if the user presses enter after typing their password. The form submit event does not appear to be triggered. Alternatively we could listen on keypress for Enter and listen on passwordNext for click.

Update the logic for which pages under accounts.google.com indicate a correct password and which indicate an incorrect password.

If a possible password is too old, then delete it. This is probably not much an issue, but I want to avoid any issues around tabIds being reused.

Update the change password page to the new URL

----

save Date (not number) in passwordalert.background.rateLimitResetDate_

This JsDoc warning actually found a bug:
background.js:903: WARNING - assignment to property rateLimitResetDate_ of passwordalert.background
found   : number
required: (Date|null)
    passwordalert.background.rateLimitResetDate_ =
    ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^

Date.protogype.setHours() modifies the instance of Date and returns milliseconds, not the Date instance: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Date/setHours

I don't think this bug actually had any effects, but this is more correct.