Fix handling of empty element in whitelist domains.  Basically the problem is that when the user does not specify anything, the policy is [""] which matches everything with endsWith().
Electing to filter it out in the content script, rather than when the user
downloads the json config since we might want to guard against manually created json
configs too.

I also test the case with comma-only, but can not repro the customer issue.
But in testing now, the comman-only is handled correctly, i.e. pwa will be in
running state.